Created useSticky hook
